requirements.txt
autopep8==1.5.3
backcall==0.2.0
cachetools==4.1.0
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.18.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.9
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
jedi==0.17.1
jupyter-client==6.1.3
jupyter-core==4.6.3
numpy==1.19.0
oauthlib==3.1.0
pandas==1.0.5
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
pip-autoremove==0.9.1
prompt-toolkit==3.0.5
protobuf==3.12.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
Pygments==2.6.1
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
spotipy==2.12.0
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.2.4
